Adds support for deep link auth
This adds support for deep link authorization by way of
levelup-deeplinkauth-lib. This is a lightweight Android project that can
be included by an app. It has an integrator which will verify the
authenticity of the LevelUp app and make the appropriate Intent request
to make the permissions request and return an access token.

* Adds levelup-deeplinkauth-lib
* Adds App/Permission/PermissionsRequest models